% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dark_kinase_list}
\alias{dark_kinase_list}
\title{Human Dark Kinases}
\format{A data frame with 162 rows and 1 column
\describe{
  \item{hgnc_id}{The human dark kinase name}
}}
\source{
NIH and the Dark Kinase DRGC
}
\usage{
dark_kinase_list
}
\description{
A dataset containing a list of dark kinases
}
\keyword{datasets}
