% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarkKinaseTools.R
\docType{package}
\name{DarkKinaseTools}
\alias{DarkKinaseTools}
\alias{DarkKinaseTools-package}
\title{DarkKinaseTools: A package for working with dark kinase data.}
\description{
The DarkKinaseTools package provides three data sets that I've found to be
generally usefull for dealing with kinase data sets. These include a set of
all human kinase, all dark human kinases and a set of mouse kinases and
corresponding human kinases.
}
\section{Dataset}{
NA
}

