% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinase_props.R
\name{get_dark_kinase_names}
\alias{get_dark_kinase_names}
\title{Get Kinase Lists}
\usage{
get_dark_kinase_names()
}
\value{
A data.frame with a list of the dark kinases
}
\description{
This function reads a CSV file with the list of dark kinases
}
