% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kin_domain_locations}
\alias{kin_domain_locations}
\title{Human Kinase Domain Locations}
\format{A data frame with 516 rows and 3 columns
\describe{
  \item{kinase_com_name}{The name used on the kinase.com list, used for
                         cross-referencing other lists from kinase.com}
  \item{start}{First amino acid in the kinase domain}
  \item{stop}{Last amino acid in the kinase domain}
}}
\source{
\url{http://kinase.com/}
}
\usage{
kin_domain_locations
}
\description{
A dataset containing a list of human kinase domain start and stop locations.
The start and stop locations are coded by the position within the kinase, so
a 250 amino acid kinase domain might start at position 200 and continue to
position 450.
}
\keyword{datasets}
