% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_human_gene_names}
\alias{mouse_human_gene_names}
\title{Mouse Kinases and Corresponding Human Orthologs}
\format{A data frame with 463 rows and 2 columns
\describe{
  \item{MGI_symbol}{The kinase symbol from the MGI database}
  \item{kinase_com_human_orthlog}{The corresponding human ortholog from kinase.com if identified}
}}
\source{
\url{http://kinase.com/mouse/} and the MGI database
}
\usage{
mouse_human_gene_names
}
\description{
A dataset containing a list of mouse kinases and human orthologs, if present.
This list is mostly from the kinase.com website with the pseudogenes removed
and the mouse gene names updated to match the current listing in the Jackson
Labs MGI data.
}
\keyword{datasets}
