% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_kinases}
\alias{all_kinases}
\title{All Human Kinases}
\format{A data frame with 524 rows and 5 columns
\describe{
  \item{hgnc_id}{The human dark kinase ID from HGNC}
  \item{symbol}{The human dark kinase symbol from HGNC}
  \item{ensembl_gene_id}{ID from ensembl}
  \item{class}{Either Dark or Light, depending on the understudied-ness}
  \item{kinase_com_name}{The name used on the kinase.com list, used for cross-referencing other lists from kinase.com}
}}
\source{
NIH, the Dark Kinase DRGC, HGNC and kinase.com
}
\usage{
all_kinases
}
\description{
A dataset containing a list of human kinases. This list has several
pseudokinases, as they are included in the dark kinase list.
}
\keyword{datasets}
