% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dark_kinases}
\alias{dark_kinases}
\title{Human Dark Kinases}
\format{A data frame with 162 rows and 2 columns
\describe{
  \item{hgnc_symbol}{The human dark kinase name in HGNC}
  \item{DRGC_symbol}{The symbol used int the DRGC documents, mostly idenfical to the hgnc_symbol, minus proteins without HGNC symbols}
}}
\source{
NIH and the Dark Kinase DRGC
}
\usage{
dark_kinases
}
\description{
A dataset containing a list of dark kinases.
}
\keyword{datasets}
