% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarkKinaseTools.R
\docType{package}
\name{DarkKinaseTools}
\alias{DarkKinaseTools}
\alias{DarkKinaseTools-package}
\title{DarkKinaseTools: A package for working with dark kinase data.}
\description{
The DarkKinaseTools package provides three data sets that I've found to be
generally usefull for dealing with kinase data sets. These include a set of
all human kinase, all dark human kinases and a set of mouse kinases and
corresponding human kinases. Complete documentation for each data set is
available through help(data_set_name). Unless otherwise noted, all of these
kinase are human.
}
\section{Datasets}{


  \itemize{
    \item dark_kinases: List of Dark Kinases
    \item all_kinases: List of All Kinases
    \item mouse_human_gene_names: Mapping between human-mouse kinase homologs
    \item kin_domain_locations: Listing of kinase domain locations and protein lengths
  }
}

