% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kin_domain_locations}
\alias{kin_domain_locations}
\title{Human Kinase Domain Locations}
\format{A data frame with 516 rows and 8 columns
\describe{
  \item{kinase_com_name}{The name used on the kinase.com list, used for cross-referencing other lists from kinase.com}
  \item{start}{First amino acid in the kinase domain}
  \item{stop}{Last amino acid in the kinase domain}
  \item{protein_length}{Number of Amino Acids in this protein}
  \item{hgnc_id}{The ID number assigned by HGNC, formated as HGNC:#####}
  \item{symbol}{The human dark kinase name in HGNC}
  \item{ensembl_gene_id}{The ID assigned by ensembl, all start with ENSG}
  \item{class}{Either Light or Dark, using the IDG classifications}
}}
\source{
\url{http://kinase.com/}
}
\usage{
kin_domain_locations
}
\description{
A dataset containing a list of human kinase domain start and stop locations.
The start and stop locations are coded by the position within the kinase, so
a 250 amino acid kinase domain might start at position 200 and continue to
position 450. The undocumented columns in this data set are identical to the
data in the all_kinases data set. I found it convenient to have these
additional identifiers here as well.
}
\keyword{datasets}
