% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_human_gene_names}
\alias{mouse_human_gene_names}
\title{Mouse Kinase and Corresponding Human Orthologs}
\format{A data frame with 555 rows and 2 columns
\describe{
  \item{Mouse_gene_name}{The mouse kinase name}
  \item{Human_ortholog}{The corresponding human ortholog gene name, if identified}
}}
\source{
\url{http://kinase.com/mouse/}
}
\usage{
mouse_human_gene_names
}
\description{
A dataset containing a list of mouse kinases and human orthologs, if present.
This list is directly from the kinase.com website with the pseudogenes remove
and most of the columns.
}
\keyword{datasets}
